Thread Creation and execution procedures cleanup
Procedures Create and Execute in M & J threads definitions had assignments from objects from the main thread. This worked because of direct visibility, but it was a conceptual error.
Only Synchronization procedures (which are executed in the main thread) may contain such references, but this also means a separate unit with the objects cannot be created because the visibility disappears. Call back functions or some other way, but I need to search more for solutions.